code refactoring; users cannot be duplicated
