+++ backend/src/controllers/DevController.js
@@ -1,0 +1,27 @@
+const axios = require('axios')
+const Dev = require('../models/Dev')
+
+module.exports = {
+    async store (req, res) {
+        const { github_username, techs } = req.body
+
+        let dev = await Dev.findOne({ github_username })
+    
+        if(!dev) {
+            const response = await axios.get(`https://api.github.com/users/${ github_username }`)
+            const { name = login, avatar_url, bio } = response.data
+        
+            const techsArray = techs.split(',').map(tech => tech.trim())
+        
+            dev = await Dev.create({
+                github_username,
+                name,
+                avatar_url,
+                bio,
+                techs: techsArray
+            })
+        }
+    
+        return res.json(dev)
+    }
+}
