+++ backend/src/routes.js
@@ -1,6 +1,5 @@
 const { Router } = require('express')
-const axios = require('axios')
-const Dev = require('./models/Dev')
+const DevController = require('./controllers/DevController')
 
 const routes = Router()
 
@@ -8,23 +7,6 @@ routes.get('/', (req, res) => {
     res.json({ message: 'Hello World' })
 })
 
-routes.post('/devs', async (req, res) => {
-    const { github_username, techs } = req.body
-
-    const response = await axios.get(`https://api.github.com/users/${ github_username }`)
-    const { name = login, avatar_url, bio } = response.data
-
-    const techsArray = techs.split(',').map(tech => tech.trim())
-
-    const dev = await Dev.create({
-        github_username,
-        name,
-        avatar_url,
-        bio,
-        techs: techsArray
-    })
-
-    return res.json(dev)
-})
+routes.post('/devs', DevController.store)
 
 module.exports = routes
